+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
-
+- Add [`no-relative-parent-imports`] rule: disallow relative imports from parent directories.
 
 ## [2.12.0] - 2018-05-17
 ### Added